Update README.md
Lesson #9 has an error in the Jest Test code given to the student. The code is importing the function AFTER the test is initialized and this causes the test to fail with:

TypeError: fromEuroToDollar is not a function.

So I moved the import to be outside of the test. We can verify this by running the test before and after. Also, back in lesson #6, it was done the correct way 💪